RankingsListType belongs to the provider
